+++ keramics-formats/src/ntfs/index.rs
@@ -157,7 +157,7 @@ impl NtfsIndex {
                 );
                 self.block_ranges.push(block_range);
 
-                virtual_cluster_number += data_run.number_of_blocks as u64;
+                virtual_cluster_number += data_run.number_of_blocks;
                 virtual_cluster_offset += range_size;
             }
             if cluster_group.last_vcn != 0xffffffffffffffff
